got the Voice chat streaming working finally
i added logging as well

i am now trying to setup the radio-browser commands. once i get them working and able to produce a valid url to use with the play_url command i will try to make new ones to just start playing the first (or a random) entry in the results.